Remove Freifunk Vogtland cities with 0 nodes
The definition of a location/city in the directory for a supra-regional
community is not well defined. But location/cities without a node seem to
be disfavored and should be part of this directory.

Bad Elster, Heinsdorfergrund, Klingenthal and Neumark have no nodes since
at least 14 days. The Freifunk Vogtland community must therefore be assumed
not existing in these locations.